Signed in saved case confirmation email.
Previously, when a user is already signed in, any case they create was being
assigned to the user without triggering a 'saved case' confirmation email.

With this change we ensure the assigning of the user is also triggering the
confirmation email, making use of the service object `SaveCaseForLater`.

Only one email will be sent as once the case is assigned to the signed in user,
no more emails are sent.

https://www.pivotaltracker.com/story/show/144695107