Refactor the landing page step for appeal and closure.
This PR takes out of the steps hierarchy the landing pages (start pages) for appeal
and closure in order to mitigate an issue with having to create a tribunal case too
early in the journey, before even the user answer any question at all.

Until now it was not very important, but with save & return we don't want to assign
a tribunal case to a user (and this send a confirmation email) until the user really
advance a bit more in the journey and answer the first question (case type).

Please note, in case it is important for Google Analytics, the following URLs have changed:

* /steps/appeal/start  -> /appeal
* /steps/closure/start -> /closure

This work is related to a previous PR:
https://www.pivotaltracker.com/story/show/144695107