Validate blank email addresses on reset password.
Devise Paranoid mode means it considers any email (even blank ones) as valid and
will not disclose any information regarding if it is valid, exists or not.
This is cool, but at least blank emails should be shown as errors to the user.

This PR intercept the errors and if a :blank email error is set, we re-render the
view in order to show the errors, otherwise normal flow happens.
This will however not show errors for malformed email addresses as Devise will use
this code also when the email doesn't exists, which means we would be disclosing
valid, existing accounts.

If we wanted to validate the email address some extra monkey patching might be needed.

https://www.pivotaltracker.com/story/show/144345627