Replacing common invalid hyphens and quotes from email user input.
Due to copy&paste or whatever other reason, some users might be entering invalid
characters into the email fields, thus making the validation fail. It is very difficult
to catch some of these characters as they are visually equivalent, so the user will get
frustrated thinking their email address is invalid.

This PR solves (to a certain extent) the problem by replacing silently some of the most
common equivalent characters (for hyphens and single quotes) with the valid ones.

https://www.pivotaltracker.com/story/show/144937767