Refactor the 'case not found' error into 2 separate errors.
This PR introduces a small refactor to separate into 2 different errors the
`case not found` error, as we now need to show a more accurate message when
the case doesn't exist in the database as opossed to the session.

The `case not found` error is now raised on the cases controller when the
record does not exist, and the `invalid session` error is raised in any other
place where a valid tribunal case is needed to be in the session.

https://www.pivotaltracker.com/story/show/145025665