Create a management report that emails stats daily (#448)
* Fix odd spec failures

David S. was getting consistent failures in these specs, due to
TribunalCase ids not being as expected. Making this change prevents
these errors. So, adding this commit, even though it shouldn't be
necessary.

* add a NotifyMailer#statistics_report method

This sends any string data to the configured statistics report
recipient email address (taken from an env var), with the
supplied title as the email subject

* Add TaxTribs::StatisticsReport service class

* mutation test all TaxTribs::* classes

* email a statistics report every day

* add explicit tests for mailer env vars

These tests confirm that we really are interrogating the
environment, to determine these values