use 'deliver!' to send email from a rake task (#457)
* use 'deliver!' to send email from a rake task

deliver_later works fine inline in the rails application code,
but it doesn't work for rake tasks, so the daily statistics
email was not being sent. This commit fixes that.

* replace explicit ordering of statistics_report results

I removed the "ORDER BY" clause, because mutant was bitching
about it. But, it turns out I just didn't have enough records
for the ORDER BY to always be necessary