Add unique page titles to every page, and fallback mechanism.
This PR adds unique page titles to every single page in our service, and put
in place a mechanism to alert us in case a page title is missing without raising
an error to the user, and defaulting to the 'base' title.

This should require a revision by Josh to ensure all the titles are fine. I've
gone for sensible defaults as a first version.

https://www.pivotaltracker.com/story/show/145128463