Add new challenge radio option for direct tax and restoration
As a user (direct tax or restoration) I want to inform the tribunal that I have
appealed to HRMC but was refused due to appealing late.

Selecting this option the user proceeds as normal to dispute type.

Two separate values have been added to the ChallengedDecisionStatus value object
so the copy can be different depending on the case type being direct tax or a
restoration case (this was required by Josh).

https://www.pivotaltracker.com/story/show/145026639