Add fieldset legends with visually hidden class.
As part of the DAC accessibility report, an issue was raised to add fieldset legends,
but we wanted those to be visually hidden, which `govuk_elements_form_builder` gem doesn't
support in their current version.

A PR has been submitted to `govuk_elements_form_builder` to support adding custom html
attributes to the inner element inside the legend so we can add `class="visuallyhidden"`

https://github.com/ministryofjustice/govuk_elements_form_builder/pull/89

For now we are locking the Gemfile to the PR branch, and once it is merged we will get back
to whatever version it is (probably v0.0.6).

Note: some other gems in the Gemfile has been bumped to minor releases as part of updating
the gems `govuk_elements_rails` and `govuk_frontend_toolkit` in preparation for the next
major releases of these gems (currently with breaking changes so we can't update just yet).

https://www.pivotaltracker.com/story/show/145268263