Customise password label (and remove hint) from signin form
Because the way active record attributes are translated, the User model will have
their `email` and `password` fields always show the same labels (and hint) in every
form, regardless if it is the sign-in or the sign-up, and this is confusing.

In order to customise these labels and remove the hint *only* for the sign-in form,
we need to override some default values. I've added a comment as this is not very
intuitive on first sight.

https://www.pivotaltracker.com/story/show/145424145