Add title and language to the generated case PDF
This was raised as an issue in the DAC report. As we can't figure out how to
add a 'language' attribute, we've added in the meantime the language in the
title itself, as we have full control of this attribute.

`Title (English)`

If this is not satisfactory we will figure out any other solution.

This work was related to this ticket and needs to be done also in the pro-forma:
https://www.pivotaltracker.com/story/show/145201729