Try for specific error, fallback to general error
Locally, I've seen both the format error and the server code error.  I
am uncertain *which* server and whether there might be a range of
possible errors.  To address this, I am taking the double-chekcing
approach on dropping files-look for an error on the drop, and check for
the No files error.

DS and I discussed the possiblity of uploading in another part of the
app.  However, I must point out that it already does this with the
ground for appeal document.  It also checks to ensure the virus scanning
is working.