Implement Zendesk feedback form (#494)
This is a first version of a feedback form submitting to Zendesk. The code is
flexible enough to be extended to implement other kind of forms if we decided so
(i.e. bug report) and also to implement the `end of journey` survey, replicating
as much as possible the current Google Forms one.

Two new ENV variables are needed. If they are not set, the app will not fail but
will give an exception when trying to submit the feedback to Zendesk:

* ZENDESK_USERNAME
* ZENDESK_TOKEN

https://www.pivotaltracker.com/story/show/144703859