Changes to feedback form and header link
Some copy changes to the feedback form, adding an optional email field, and
removing the rating radios, as per Alejandra instructions.

The rating radios will be only in the end of journey surveys.

The form is now accessible through the link in the header but will not work
unless the ENV variables exists.