Tidy production logging (#502)
* Fix local rspec failure

The issue described in the comment is only happening on my machine.
I am putting this workaround in place as I am otherwise unable to
progress.

* Make the logs easier to read

Turn off shell colorizing and get rid of action_view timings in
production.

* Prevent leakage of sensitive information in logs

* Enable tracking of successful uploads

We can see successful uploads on the uploader, but have not previously
been logging the attempts when they start here. This corrects that. It
will log the details of the call *before* it is made (and for deletes).
A call that does not raise an error is successful.