Enable non-JS users to reach post-service survey
This commit gets around a problem where IE8, for which we disable
javascript, does not get redirected to our survey, because the
'Finish' button uses javascript to send a DELETE request when the
button is clicked. This commit makes the sessions controller
handle GET and DELETE requests in the same way.