Enable tracing of docker container serving request (#518)
Prior to this, the `host` attribute was populated with a hash that
identified the specific docker container responding to the request. This
is far more useful than the service URL as it enables us to connect the
sentry error reports to the log file entries.

This commit returns us to using the docker container hash.