Dynamic app root based on requested domain
As part of the preparations for the going live with a GDS domain, and a
gds-hosted homepage, we need to be able to handle during some time access
to the app through both URLs/domains but there are some small differences:

* If accessing through our current dsd.io domain, everything will continue
as it is now, and the `/` will take to our homepage with a button to `/start`

* If accessing through a gov.uk domain, our `/` will redirect to `/start` as
GDS will be hosting the homepage with the `start` button for us, and thus we
just need to show users coming through this URL the task list.

Temporarily this will be a redirect, but once the GDS domain is live and
working we will be able to move our `/start` to be our `/`.

* Another change will be for the link in the header to direct users to our
`/` if accessing through a dsd.io domain or to the TBC GDS service homepage
if accessing through a gov.uk domain.

This is a part of a bigger ticket:
https://www.pivotaltracker.com/story/show/146376375