Stop replacing hyphens with `&dash;` on save
We run into a pesky problem where we thought the NormalisedEmailType was not replacing
properly some unicode hyphens not valid for use in email addresses, or was doing so but
not persisting the replaced (and valid) character to the database.

It turned out the problem was in another different part of the app, the sanitizing of
the database text/string fields `on save`.

The way this bug was reproducing is:

1) The user entered their email address containing an invalid unicode hyphen (for all we
know introducing a valid hyphen would also trigger this bug but for the sake of this
explanation I will go with an invalid hyphen).

2) Our `NormalisedEmailType` would silently replace the invalid hyphen with their valid
equivalent `-` so the user don't get an 'invalid email address' error on validation.

3) On trying to persist this new valid email address to the database, there is a callback
to run TribunalCase#sanitize through the database fields, and sadly enough, the (now valid)
hyphen would get converted back to an html representation of it `&dash;` making the email
address invalid again and persisting this invalid one to database (with char ord code 8208
instead of valid one 45).

4) This email will get all the way to GLiMR (they might run into problems because of this)
but on trying to send the confirmation email to the user, Notify will reject the email as
being not valid (because the invalid hyphen).

5) If the user were to use S&R or go back to the form, they will see a valid email address
as the NormalisedEmailType would replace back the invalid char to a valid one, but on save,
we go back to point (3) and we end up again with an invalid hyphen in the database.

6) All this really happen regardless if the user enters a valid or an invalid hyphen.