Remove fake homepage and move `start` to the root (#542)
We don't need anymore a service homepage with information on how to use the service,
because GDS is hosting this homepage for us. The `start` button thus, is on GDS side,
and our service root should now be the task list.

This PR cleanups all fake homepage leftovers, move our task list to be the root, and
also setup redirections so users trying the (now non-existing) `/start` path, will get
redirected to the root `/`

Also, a new set of redirections has been setup to redirect any request to the old (prod)
domain to the new domain. These requests will maintain the whole path so requests to
resume a case pointing to the old domain will redirect to the new domain resume case.

https://www.pivotaltracker.com/story/show/147418015