Enable GA ecommerce transaction tracking
It turned out the way we were tracking submissions and other events (with Goals)
is not the recommended way to track these, as it will only track 1 event per session,
so for example an accountancy firm submitting several cases in one go, will only
track as 1 case submitted, even if they submitted 10.

It is not a mayor issue as we are also tracking most of these events at the database
level and sending reporting emails, but as part of a proof of concept, let's see how
the transaction tracking goes, and even if it helps with tracking other events easily.
I've added a few to start with (most of them what we had as Goals before).

Geckoboard also support transactions/products analytics so it should be easy to start
reporting these.

https://developers.google.com/analytics/devguides/collection/analyticsjs/ecommerce