Include case type as SKU in transaction details
Further to the previous PR to track transactions, we can extend the details
of each of those tracked events by adding a SKU or item code, and this is
perfect to track the `case type` for each of these transactions.

So for example, case submissions can be analyzed by case type (VAT, income tax
and so on) and grouped together to enable better insights.