Ensure we track registrations (we need a session)
After a bit if digging, I realised we were not tracking case saved via registrations
because by the time we render the template to track the transaction, we don't have
a tribunal case in the session anymore (we were resetting it on `sign_up` method).

With this bit of refactor, we ensure we maintain the session long enough for the
template to be rendered, and reset it with an after_action filter.
Other than this, nothing changes.