Change smoketest targeting (#553)
* Preparation for shutting down the dev environment

This sets `staging` to be the target environment for the smoke tests.

* Ensure ease-of-maintenance of the CI smoketests

While it has always been straightforward to change the somketest target
environment, it required a mid-level knowledge of both the tests and the
various `docker` commands.  This change ensures everything is clearly
documented for future maintainers.

It also includes additional information concerning the rationale(s)
behind the decisions that were made to containerise the tests, and not
to run the tests directly against production.