Report page of problems uploading documents
This is a simple reporting page, listing the problems users had uploading
documents when they select said checkbox. This is what we call a postal
route as users can't continue to submit the case to GLiMR and we instruct
them to send the documentation by post instead.

The page is protected by `http basic auth`, and the user and digest of the
password (SHA256) are set as ENV variables:

* `UPLOAD_PROBLEMS_REPORT_AUTH_USER`
* `UPLOAD_PROBLEMS_REPORT_AUTH_DIGEST`

https://www.pivotaltracker.com/story/show/147615635