Add extra copy around single file uploaders
In order to minimize the confusion some users experience, where after
selecting a file they were expecting it to upload immediately, we've
added a bit of extra copy to explain the file will upload when they
click 'Continue'.

Also, removed some extra copy in the lateness reasons step as it is
not really necessary.

https://www.pivotaltracker.com/story/show/146507353