Add new letter upload step form and controller
This PR will introduce a new, simplified, letter upload to replace the current
documents checklist (dropzone uploader).

At the moment it is not wired up to the decision tree and thus is not visible to
the user, nor there are breaking changes with our current implementation, but
because the final PR might otherwise be too big, I decided to create at least 2
smaller PRs, being this one the first of those, to be easier to review.

The new form is very similar to the other steps we have where only one document
can be uploaded (i.e. `grounds for appeal`) with the difference it will maintain
the ability to check `having problems uploading` and go to the postal kick-out.
In summary, we remove the dropzone and replace it with a single file uploader,
and customise the view to talk about the specific letter needed to be uploaded.

https://www.pivotaltracker.com/story/show/148420071