Replace dropzone upload step with letter upload
This PR is a follow up of the previous one, removing all the steps,
form objects, views, etc. related to the previous document uploader
(dropzone) and changing the decision tree to make use of the new
individual letter upload step.

Also the smoke tests have been updated, but because now in the appeal
journey we don't have any step with dropzone, it felt like a good
opportunity to introduce a very simple `closure` happy path, which
will continue using the dropzone to upload supporting documents.

There is still one remaining thing (another PR to follow) to consider:
the cleaning of the (now unused) two checkboxes DB attributes, but as
there might be draft cases making use of these checkboxes or with files
uploaded, we need to create a maintenance migration to remove the
attributes but also to ensure no cases are left in a broken state.

https://www.pivotaltracker.com/story/show/148420071