Migration to remove unused letter attributes
This is the last bit of the work needed to change the dropzone step to
a single file upload for appeals.

This will remove the now unused boolean attributes, previously used to
indicate which kind of letter the user had uploaded, with the caveat there
might be draft cases that made it as far as the old `documents checklist` step.
The migration will take care of it.

https://www.pivotaltracker.com/story/show/148420071