Add new admin report page for 'other' case types
In an identical way to what we did for listing the problems users had
uploading documents, a new page have been added to list the value users
entered when selecting `none of the above` in the list of case types.

Also we've added the column `user type` to the upload problems report, as
well as using this column in this report too.

https://www.pivotaltracker.com/story/show/148836409