Fix a few issues with filenames in the uploader
Some problems were detected in the uploader, where some characters were
failing to upload or to be removed later on due to different encodings.

This PR ensures, in first place, we do not allow characters other than
valid ASCII in the filenames, and give a meaninful error to the user,
instead of uploading and silently modify their encoding and do not allow
the removal later on due to mismatches in the encodings.

Also, some characters like brackets are totally fine but by default
`URI.encode()` will not encode these characters, making our request to
delete a file containing square brackets fail. This has also been solved
by passing an extra argument to the encode() method to force encoding of
all non-word characters, which in turn will encode `[` and `]` properly.

https://www.pivotaltracker.com/story/show/149083759