New admin report for other dispute types (#567)
* Add new admin report for other dispute types

New controller and view, together with associated routing and i18n.

* Extract duplicated admin tests to shared examples

Now we have 3 identical password-protected controllers with an #index
action, and we can extract some duplicated tests to a shared file.

* Rename i18n key to match their attribute

We were previously using `updated_at` attribute in the report results but
now we are using `created_at`. This will update the i18n to be in sync and
there is no confusion.

* Rename partial to remove redundant prefix

The partial is already in the `admin` namespace so we don't really need
to have this as a prefix too.