Copy changes to include Welsh taxes
The Welsh Revenue Authority is using devolved power to take over
the collection of certain taxes from HMRC in Wales.
So in effect they are different names in Wales for two of the
existing taxes that previously HMRC collected nationally.

This has no effect on GLiMR submissions or PDF form.